% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Py_QTL_Parser.R
\name{Py_QTL_Parser}
\alias{Py_QTL_Parser}
\title{Py_QTL_Parser}
\usage{
Py_QTL_Parser(vcf, HighBulk, LowBulk)
}
\arguments{
\item{vcf}{A vcf file that is as a tidy data frame}

\item{HighBulk}{The name of the HighBulk sample}

\item{LowBulk}{The name of the LowBulk sample}
}
\description{
Parses a vcf file which has been converted to a tidy data frame before running the operation
}
